+++ storage/innobase/trx/trx0i_s.cc
@@ -718,16 +718,16 @@ static bool fill_locks_row(
 		&& (lock->type_mode & LOCK_GAP);
 	switch (lock->type_mode & LOCK_MODE_MASK) {
 	case LOCK_S:
-		row->lock_mode = 1 + is_gap_lock;
+		row->lock_mode = uint8_t(1 + is_gap_lock);
 		break;
 	case LOCK_X:
-		row->lock_mode = 3 + is_gap_lock;
+		row->lock_mode = uint8_t(3 + is_gap_lock);
 		break;
 	case LOCK_IS:
-		row->lock_mode = 5 + is_gap_lock;
+		row->lock_mode = uint8_t(5 + is_gap_lock);
 		break;
 	case LOCK_IX:
-		row->lock_mode = 7 + is_gap_lock;
+		row->lock_mode = uint8_t(7 + is_gap_lock);
 		break;
 	case LOCK_AUTO_INC:
 		row->lock_mode = 9;
